requirements.txt

django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
python-decouple
dj-database-url
black
Pillow